fix: Setup the proper routes and sync shelfs from search
From the code review comments I had to create the '/search' route and
remove the '/books' route. I also had to make sure the books response
from the search API books's shelfs are in sync with the application's
shelf. I had put the state of the shelf and some methods at the higher
level to accomplish this so that all components knew about it.